Update up.txt
The university changed the name this year and i have email adress which is refers to the new name of the university and i can't get JetBrains products for learning. I add also the university url. The url is old but you can see in left top new name of the place. 
https://www.up.krakow.pl/